# docs/openapi/leave-attendance.yaml
openapi: 3.0.3
info:
  title: IT Center Leave & Attendance API
  version: 3.0.0
  description: |
    Serverless leave, attendance, integration, and reporting endpoints running on API Gateway → Lambda → PostgreSQL.
servers:
  - url: https://{apiId}.execute-api.ap-southeast-2.amazonaws.com
    description: AWS API Gateway invoke URL
    variables:
      apiId:
        default: dev-placeholder
paths:
...
  /api/v1/attendance:
    get:
      summary: List attendance logs
      tags: [Attendance]
      operationId: listAttendance
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          description: Optional user filter (ADMIN only)
          schema:
            type: integer
        - name: from
          in: query
          description: Start timestamp (ISO 8601)
          schema:
            type: string
        - name: to
          in: query
          description: End timestamp (ISO 8601)
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: sort
          in: query
          schema:
            type: string
            default: clock_in,desc
      responses:
        '200':
          description: Paginated attendance logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAttendanceLogs'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/attendance/clock-in:
    post:
      summary: Clock in
      tags: [Attendance]
      operationId: clockIn
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockInRequest'
      responses:
        '201':
          description: Clock-in recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceLog'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '409':
          description: Active session already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/attendance/clock-out:
    post:
      summary: Clock out
      tags: [Attendance]
      operationId: clockOut
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockOutRequest'
      responses:
        '200':
          description: Clock-out recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceLog'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          description: No active clock-in session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AttendanceLog:
      type: object
      properties:
        logId:
          type: integer
        userId:
          type: integer
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: string
          format: date-time
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        source:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaginatedAttendanceLogs:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceLog'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
    ClockInRequest:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        source:
          type: string
    ClockOutRequest:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        requestId:
          type: string
          description: AWS request identifier
  responses:
    Unauthenticated:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []